+++ src/routes/(app)/admin/events/[id]/registrations/data-table-actions.svelte
@@ -23,7 +23,7 @@
 <DropdownMenu.Root>
 	<DropdownMenu.Trigger asChild let:builder>
 		<Button variant="ghost" builders={[builder]} size="icon" class="relative h-8 w-8 p-0">
-			<span class="sr-only">{$_("admin-pages.events.event-registrations.data-table.actions.open-menu")}</span>
+			<span class="sr-only">{$_("common.open-menu")}</span>
 			<Ellipsis class="h-4 w-4" />
 		</Button>
 	</DropdownMenu.Trigger>
