+++ src/routes/(app)/admin/organizations/[id]/+page.svelte
@@ -9,12 +9,19 @@
 	import { fade } from 'svelte/transition';
 	import type { PageServerData } from './$types';
 	import { _ } from '@services';
+	import { writable } from 'svelte/store';
+	import { setContext } from 'svelte';
 
 	export let data: PageServerData;
-	let organization = data?.organization;
-	let orgMembers = data?.orgMembers || [];
-	let portraitTemplates = data?.portraitTemplates || [];
-	let eventRegistrations = data?.eventRegistrations || [];
+
+	$: ({ organization, orgMembers, createInviteForm, portraitTemplates, eventRegistrations } = data);
+
+	let createInviteFormStore = writable(createInviteForm);
+	$: {
+		createInviteFormStore.set(createInviteForm);
+	}
+
+	setContext('createInviteForm', createInviteFormStore);
 
 	let activeTab = 0;
 	const tabHeadings = ['companyInformation', 'portraits', 'bookings'];
@@ -31,31 +38,27 @@
 		return keyValid && valueValid;
 	}
 </script>
-<div>
 
-	<header class="space-y-4">
-		<Link href="/admin/organizations">{$_("admin-pages.organizations.back-to-overview")}</Link>
-		<h1 class=" text-stone-950 text-3xl font-extrabold">{organization?.name}</h1>
-		<p>{$_(`common.org-types.${organization?.organizationType}`)}</p>
-	</header>
-	<div class="grid grid-cols-1 gap-4 w-full">
-		<Tabs hasBorder={false} {tabHeadings} {activeTab} {handleTabChange} />
-	</div>
-	<section class=" mt-10 max-h-[590px] overflow-y-scroll">
-		{#if activeTab === 0}
-			<div>
-				<CompanyInformationTab
-					organizationInfo={Object.entries(organization || {}).filter(filterOrgEntries).map(([key, value]) => ({ label: key, value }))}
-					{orgMembers} />
-			</div>
-		{:else if activeTab === 1}
-			<div in:fade>
-				<PortraitsTab {portraitTemplates} />
-			</div>
-		{:else if activeTab === 2}
-			<div in:fade>
-				<BookingsTab {eventRegistrations} />
-			</div>
-		{/if}
-	</section>
+<header class="space-y-4">
+	<Link href="/admin/organizations">{$_("admin-pages.organizations.back-to-overview")}</Link>
+	<h1 class=" text-stone-950 text-3xl font-extrabold">{organization?.name}</h1>
+	<p>{$_(`common.org-types.${organization?.organizationType}`)}</p>
+</header>
+<div class="grid grid-cols-1 gap-4 w-full">
+	<Tabs hasBorder={false} {tabHeadings} {activeTab} {handleTabChange} />
 </div>
+<section class=" mt-10 ">
+	{#if activeTab === 0}
+		<CompanyInformationTab
+			organizationInfo={Object.entries(organization || {}).filter(filterOrgEntries).map(([key, value]) => ({ label: key, value }))}
+			{orgMembers} />
+	{:else if activeTab === 1}
+		<div in:fade>
+			<PortraitsTab {portraitTemplates} />
+		</div>
+	{:else if activeTab === 2}
+		<div in:fade>
+			<BookingsTab {eventRegistrations} />
+		</div>
+	{/if}
+</section>
