+++ static/lang/en.json
@@ -119,7 +119,10 @@
       "nonprofit": "Nonprofit",
       "company": "Company",
       "other": "Organization"
-    }
+    },
+    "next": "Next",
+    "previous": "Previous",
+    "open-menu": "Open menu\u2026"
   },
   "components": {
     "addonPackage": {
@@ -437,7 +440,8 @@
       "upcomingEvents": "Upcoming Events",
       "registeredEvents": "Registered Events",
       "noEventsRegistered": "No events registered yet!",
-      "noEventsRegisteredDetail": "There are no events to display yet. Please register a event for your company"
+      "noEventsRegisteredDetail": "There are no events to display yet. Please register a event for your company",
+      "alreadyRegistered": "Registered"
     },
     "organizations": {
       "myOrganizations": "My Organizations",
@@ -488,7 +492,8 @@
           "fullName": "Your full name",
           "phoneNumber": "Your phone number (optional)",
           "email": "Your Work Email",
-          "responsibility": "Your responsibility"
+          "responsibility": "Your responsibility",
+          "title": "Your Salutation"
         },
         "placeholders": {
           "fullName": "Alice Anderson",
@@ -527,7 +532,6 @@
         "none-found": "No Event Registrations for this company where found",
         "": ""
       },
-      "openMenu": "Open menu\u2026",
       "data-table": {
         "actions": {
           "view-details": "View details\u2026",
@@ -644,7 +648,6 @@
             "review-catalogue-data": "Katalogdaten überprüfen",
             "confirm": "Bestätigen\u2026",
             "reject": "Ablehnen\u2026",
-            "open-menu": "Open menu\u2026",
             "actions": "Actions",
             "copy-id": "Copy registration Id"
           },
